refactor(apps): split assetsafe app into specialized apps with shared common module
The monolithic assetsafe app has been split into three specialized apps:
- asset_management: handles AssetRegistration
- collateral: handles CollateralRegistration
- hire_purchase: handles HirePurchaseRegistration

Shared models and enums (PartyType, BaseAssetType, etc.) have been moved to a new
apps.common app to avoid duplication.

All models now import from apps.common.models instead of apps.assetsafe.models.
Migrations have been updated to depend on common instead of assetsafe.
The assetsafe app and its files have been removed from INSTALLED_APPS and deleted.

BREAKING CHANGE: The assetsafe app has been removed. Update any imports or references
to use the new specialized apps and apps.common.
