+++ apps/asset_management/models/__init__.py
@@ -1,4 +1,3 @@
 from .models import AssetRegistration
 
 __all__ = ["AssetRegistration"]
-
