--- apps/assetsafe/api/urls.py
@@ -1,11 +1,0 @@
-"""
-urls.py — AssetSafe (shared models only)
-
-This app now serves only as a shared models/enums repository.
-API endpoints have been moved to dedicated apps:
-- apps.asset_management (AssetRegistration)
-- apps.hire_purchase (HirePurchaseRegistration)
-- apps.collateral (CollateralRegistration)
-"""
-
-urlpatterns = []
