--- apps/assetsafe/apps.py
@@ -1,10 +1,0 @@
-"""Assetsafe API app configuration."""
-
-from django.apps import AppConfig
-
-
-class AssetsafeConfig(AppConfig):
-    """Assetsafe API app configuration."""
-
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.assetsafe"
