--- apps/assetsafe/models/__init__.py
@@ -1,17 +1,0 @@
-from .models import (
-    PartyType,
-    BaseAssetType,
-    CollateralAssetType,
-    AssetCondition,
-    Currency,
-    TimeStampedModel,
-)
-
-__all__ = [
-    "PartyType",
-    "BaseAssetType",
-    "CollateralAssetType",
-    "AssetCondition",
-    "Currency",
-    "TimeStampedModel",
-]
