+++ apps/collateral/models/__init__.py
@@ -1,4 +1,3 @@
 from .models import CollateralRegistration
 
 __all__ = ["CollateralRegistration"]
-
