+++ apps/hire_purchase/models/__init__.py
@@ -1,4 +1,3 @@
 from .models import HirePurchaseRegistration
 
 __all__ = ["HirePurchaseRegistration"]
-
