+++ apps/hire_purchase/models/models.py
@@ -10,7 +10,7 @@ from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
-from apps.assetsafe.models import (
+from apps.common.models import (
     PartyType,
     BaseAssetType,
     AssetCondition,
